- refactoring - fixed exception on empty rules in realm_access - made sure to read the error body (and throw a exception) inorder to prevent connection leaking and to improve error reporting.
