+++ settings.gradle.kts
@@ -1,3 +1,4 @@
+@file:Suppress("UnstableApiUsage")
 dependencyResolutionManagement {
     repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
     repositories {
